pyproject.toml
[tool.poetry]
name = "cogtiler"
version = "0.1.0"
description = ""
authors = ["Sybrand Strauss <sybrand.strauss@gov.bc.ca>"]

[tool.poetry.dependencies]
python = "^3.8"
rio-tiler = "^3.1.6"
fastapi = "^0.79.0"
python-decouple = "^3.6"
gunicorn = "^20.1.0"
uvicorn = "^0.18.2"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
notebook = "^6.4.12"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
